fix: disable NixOS module system checks
If a patch adds a new NixOS option and the user tries to use that in
their config, the eval would fail without this, with the error
"The option X.Y.Z does not exist."

First we evaluate the config with unpatched nixpkgs to get the system
from it (and the nixpkgs-patcher config), then we create the patched
nixpkgs which will evaluate the final config that will build the system.

Previously avoiding the first eval by the user setting the system (and
nixpkgs-patcher module disabled) was a workaround.